==> 2. Some basic career analysis (Seb Ogier).ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
from urllib.parse import urlencode
import urllib.parse
from copy import copy
import requests
import re
import json
import urllib.request

df = pd.read_csv('Sébastien_Ogier.csv') 
df

# Total Race Event Wins

df.groupby(['Name','Year']).size().reset_index(name='count')

df.loc[df['Race_Result'] == "1"].groupby(['Name','Year']).size().reset_index(name='count')

72/169

# Total Stage Wins

df.count()

df.loc[df['stage'] == 1].count()

668/3068

# How many DNF?

df.loc[df['DNG_Reason'] != "Null"].groupby(['Name','Year'])['DNG_Reason'].size().reset_index(name='count')

13/169

# Find smaller accidents and come-backs

grouped = df.groupby(['Name','Year'])
results = []

for name, group in grouped:
    group['diff'] = group['stage'].diff()
    
    for index, row in group.iterrows():
        if row['diff'] > 20 or row['diff'] < -20:
            results.append([row['Name'], row['Year'], row['ssname'], row['diff']])
            
results
