==> 3. Scrape all the race info.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
from urllib.parse import urlencode
import urllib.parse
from copy import copy
import requests
import re
import json
import urllib.request

year = []
a = 1
for x in range(2000, 2024):
    my_url = "https://www.ewrc-results.com/season/" + str(x) + "/"
    raw_html = requests.get(my_url).content
    soup_doc = BeautifulSoup(raw_html, "html.parser")
    
    events = soup_doc.find_all("div", class_= "season-event text-center mt-3 position-relative")

    valid_events = []
    for event in events:
        if not event.find("div", class_= "widget-canceled"):
            valid_events.append(event)

    events = valid_events
    
    

    for event in events:
        race = {}
        race['Serial'] = a
        race['URL'] = event.find("div", class_= "season-event-name fs-13 pt-3").find("a")['href']
        
        url = event.find("div", class_= "season-event-name fs-13 pt-3").find("a")['href']
        name = re.findall("/final/(.+)/",url)[0]
        race['Name'] = name
    
        race['Year'] = x
        year.append(race.copy())
        a = a + 1
        
    print(x)
year

df = pd.DataFrame(year)
df

df.to_csv('race_list.csv')

my_url = "https://www.ewrc-results.com/season/2010/"
raw_html = requests.get(my_url).content
soup_doc = BeautifulSoup(raw_html, "html.parser")
    
events = soup_doc.find_all("div", class_= "season-event text-center mt-3 position-relative")

valid_events = []
for event in events:
    if not event.find("div", class_= "widget-canceled"):
        valid_events.append(event)

events = valid_events
events

event = events[1]
event

url = event.find("div", class_= "season-event-name fs-13 pt-3").find("a")['href']
url

name = re.findall("/final/(.+)/",url)[0]
name
