==> 4. All careers.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
from urllib.parse import urlencode
import urllib.parse
from copy import copy
import requests
import re
import json
import urllib.request

df1 = pd.read_csv('Sébastien_Loeb.csv') 
df2 = pd.read_csv('Sébastien_Ogier.csv') 
df3 = pd.read_csv('./Other_Drivers/Carlos_Sainz.csv') 
df4 = pd.read_csv('./Other_Drivers/Colin_McRae.csv') 
df5 = pd.read_csv('./Other_Drivers/Elfyn_Evans.csv') 
df6 = pd.read_csv('./Other_Drivers/Mikko_Hirvonen.csv') 
df7 = pd.read_csv('./Other_Drivers/Ott_Tänak.csv') 
df8 = pd.read_csv('./Other_Drivers/Thierry_Neuville.csv') 
df9 = pd.read_csv('./Other_Drivers/Tommi_Mäkinen.csv') 

master_df = pd.concat([df1, df2, df3, df4, df5, df6, df7, df8, df9])

master_df.head(5)

master_df = master_df.drop(['Unnamed: 0', 'Php_url','p_overall','stage','ssname'], axis=1)
master_df.head(5)

master_df = master_df.drop_duplicates()
master_df.head(5)

def extract_race_name(url):
    match = re.search(r"/entryinfo/(.+)/.+/", url)
    if match:
        return match.group(1)
    else:
        return "Unknown"

master_df["Match"] = master_df["URL"].apply(extract_race_name)
master_df

# Call the race df

race_df = pd.read_csv('race_list.csv') 
race_df

# Merge two dfs

merged_df = master_df.merge(race_df, left_on='Match', right_on='Name', how='left')
merged_df = merged_df.sort_values(by='Serial')
merged_df

merged_df.to_csv('Merged.csv')

import pandas as pd

# Get the unique drivers in master_df
drivers = master_df['Driver'].unique()

# Create an empty data frame to store the result
result_df = pd.DataFrame()

# Loop over each driver
for driver in drivers:
    # Filter the merged_df by driver
    driver_df = merged_df[merged_df['Driver'] == driver]
    
    # Check if the driver has an entry in each race
    missing_races = set(race_df['Name']) - set(driver_df['Name'])
    
    # If there are missing races, create a row with null values
    if len(missing_races) > 0:
        null_row = pd.DataFrame({
            'Driver': [driver],
            'Name': [None],
            'URL': [None],
            'Year': [None],
            'Race_Result': [None],
            'DNG_Reason': [None],
            'Match': [None],
            'Unnamed: 0': [None],
            'Serial': [None],
            'URL': [None],
            'Year.x': [None]
        })
        result_df = pd.concat([result_df, null_row])
    
    # Add the driver's data to the result
    result_df = pd.concat([result_df, driver_df])

# Order the result by Driver and Year
result_df = result_df.sort_values(['Driver', 'Year'])

# Reset the index of the result
result_df = result_df.reset_index(drop=True)

result_df
