==> 6. Smaller Comparison Viz.ipynb <==
%load_ext rpy2.ipython
%load_ext autoreload
%autoreload 2

%matplotlib inline  
from matplotlib import rcParams
rcParams['figure.figsize'] = (16, 100)

import warnings
from rpy2.rinterface import RRuntimeWarning
warnings.filterwarnings("ignore") # Ignore all warnings
# warnings.filterwarnings("ignore", category=RRuntimeWarning) # Show some warnings

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from IPython.display import display, HTML

%%javascript
// Disable auto-scrolling
IPython.OutputArea.prototype._should_scroll = function(lines) {
    return false;
}

%%R

# My commonly used R imports

require('tidyverse')
library(ggbump)
library(grid)
library(ggplot2)

%%R
 
df <- data.frame(
  Driver = c("Sébastien Loeb", "Sébastien Ogier"),
  Total_WRC_Championship_wins = c(9,8),
  Stage_Win_Rates = c(0.31, 0.22)
)

# Illustration for career wins

%%R

# Create a data frame with x and y coordinates for each point in the grid
df <- expand.grid(x = 1:10, y = 10:1)

# Add a column to the data frame to indicate whether each point should be red or gray
df$color <- ifelse(row_number(df) <= 31, "#F2F2F2" , "#fc3b2b")

# Create the plot
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = color), color = "white") +
  scale_fill_manual(values = c("#fc3b2b", "#F2F2F2"), guide = "none") +
  coord_equal() +
  theme_void()

ggsave("sl.pdf")

%%R

# Create a data frame with x and y coordinates for each point in the grid
df <- expand.grid(x = 1:10, y = 1:10)

# Add a column to the data frame to indicate whether each point should be blue or gray
df$color <- ifelse(row_number(df) <= 22, "#F2F2F2", "#1d98ca")

# Create the plot
ggplot(df, aes(x = x, y = y, fill = color)) +
  geom_tile(color = "white") +
  scale_fill_manual(values = c("#F2F2F2", "#1d98ca")) +
  theme_void()

ggsave("so.pdf")

# Illustration for DNF

%%R

# Create a data frame with x and y coordinates for each point in the grid
df <- expand.grid(x = 1:10, y = 10:1)

# Add a column to the data frame to indicate whether each point should be red or gray
df$color <- ifelse(row_number(df) <= 13, "#F2F2F2" , "#fc3b2b")

# Create the plot
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = color), color = "white") +
  scale_fill_manual(values = c("#fc3b2b", "#F2F2F2"), guide = "none") +
  coord_equal() +
  theme_void()

ggsave("sl-dnf.pdf")

%%R

# Create a data frame with x and y coordinates for each point in the grid
df <- expand.grid(x = 1:10, y = 1:10)

# Add a column to the data frame to indicate whether each point should be blue or gray
df$color <- ifelse(row_number(df) <= 8, "#F2F2F2", "#1d98ca")

# Create the plot
ggplot(df, aes(x = x, y = y, fill = color)) +
  geom_tile(color = "white") +
  scale_fill_manual(values = c("#F2F2F2", "#1d98ca"), guide = "none") +
  theme_void()

ggsave("so-dnf.pdf")
